requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
plotly>=5.20.0
pandas>=2.0.0